clean up tests for controlled decompositions
